Simply api routes name and move them into separate newsletter dir inside pages/api, added constants file and use it store api routes
