+++ lib/constants.ts
@@ -1,0 +1,8 @@
+const constants = {
+  newsletterSubscribeApiRoute: '/api/newsletter/subscribe',
+  newsletterResendApiRoute: '/api/newsletter/resend',
+  newsletterVerifyApiRoute: '/api/newsletter/verify',
+  verifyEmailTemplatePath: '/public/emailTemplates/emailconfirmation.html',
+};
+
+export default constants;
